Move artist_payload, filter_artist_genres, filter_artist_agency, and artist_sort to ArtistUtilities module
